Refactoring and cleaning
