+++ index.html
@@ -92,7 +92,7 @@
     <script src="views/descendants/descendants.js"></script>
     <link rel="stylesheet" href="views/descendants/descendants.css" />
 
-    <script src="views/familyGroupApp/familyGroupApp.js"></script>
+    <script type="module" src="views/familyGroupApp/familyGroupApp.js"></script>
     <link rel="stylesheet" href="views/familyGroupApp/familyGroupApp.css" />
 
     <link rel="stylesheet" href="tree.css" />
